requirements/dev.txt
requests
python-telegram-bot --pre
flake8
isort
black
pre-commit